+++ RealmSwift/SwiftUI.swift
@@ -1052,6 +1052,8 @@ private class ObservableAsyncOpenStorage: ObservableObject {
     var configuration: Realm.Configuration?
     var partitionValue: AnyBSON
 
+    var setupHasRun = false
+
     // Tracks User State for App for Multi-User Support
     enum AppState {
         case loggedIn(User)
@@ -1145,7 +1147,6 @@ private class ObservableAsyncOpenStorage: ObservableObject {
 
         if let user = app.currentUser {
             appState = .loggedIn(user)
-            asyncOpenForUser(user)
         } else {
             appState = .loggedOut
             asyncOpenState = .waitingForUser
@@ -1250,7 +1251,12 @@ private class ObservableAsyncOpenStorage: ObservableObject {
 
     /// :nodoc:
     public var wrappedValue: AsyncOpenState {
-        storage.asyncOpenState
+        if !storage.setupHasRun {
+            storage.asyncOpen()
+            storage.setupHasRun = true
+        }
+        return storage.asyncOpenState
+
     }
 
     /**
@@ -1269,6 +1275,21 @@ private class ObservableAsyncOpenStorage: ObservableObject {
                  if empty the user configuration will be used.
      - parameter timeout: The maximum number of milliseconds to allow for a connection to
                  become fully established., if empty or `nil` no connection timeout is set.
+
+     - Note: Adding a configuration on this property wrapper will use the file url defined on the `Realm.Configuration` for the
+             synced realm. Keep in mind this may conflict in case you are opening different realm files with different partitions or
+             using different users if the file url is always `default.realm`.
+     - Note: It is recommend to use the configuration obtained from the user `user.configuration(partitionValue:)`
+             when adding a configuration to this property wrapper, and modify any configuration option from it.
+
+             AsyncOpenView()
+                 .environment(\.realmConfiguration, getConfiguration())
+
+             func getConfiguration() -> Realm.Configuration {
+                 var configuration = user.configuration(partitionValue: "partitionValue")
+                 configuration.encryptionKey = getKey()
+                 return configuration
+             }
      */
     public init(appId: String? = nil,
                 partitionValue: Partition,
@@ -1284,7 +1305,6 @@ private class ObservableAsyncOpenStorage: ObservableObject {
             let bsonValue = AnyBSON(partitionValue)
             if storage.partitionValue != bsonValue {
                 storage.partitionValue = bsonValue
-                storage.asyncOpen()
             }
         }
 
@@ -1295,7 +1315,6 @@ private class ObservableAsyncOpenStorage: ObservableObject {
                 storage.partitionValue = partitionValue
             }
             storage.configuration = configuration
-            storage.asyncOpen()
         }
     }
 }
@@ -1358,7 +1377,11 @@ private class ObservableAsyncOpenStorage: ObservableObject {
 
     /// :nodoc:
     public var wrappedValue: AsyncOpenState {
-        storage.asyncOpenState
+        if !storage.setupHasRun {
+            storage.asyncOpen()
+            storage.setupHasRun = true
+        }
+        return storage.asyncOpenState
     }
 
     /**
@@ -1377,6 +1400,21 @@ private class ObservableAsyncOpenStorage: ObservableObject {
                  if empty the user configuration will be used.
      - parameter timeout: The maximum number of milliseconds to allow for a connection to
                  become fully established, if empty or `nil` no connection timeout is set.
+
+     - Note: Adding a configuration on this property wrapper will use the file url defined on the `Realm.Configuration` for the
+             synced realm. Keep in mind this may conflict in case you are opening different realm files with different partitions or
+             using different users if the file url is always `default.realm`.
+     - Note: It is recommend to use the configuration obtained from the user `user.configuration(partitionValue:)`
+             when adding a configuration to this property wrapper, and modify any configuration option from it.
+
+             AsyncOpenView()
+                 .environment(\.realmConfiguration, getConfiguration())
+
+             func getConfiguration() -> Realm.Configuration {
+                 var configuration = user.configuration(partitionValue: "partitionValue")
+                 configuration.encryptionKey = getKey()
+                 return configuration
+             }
      */
     public init(appId: String? = nil,
                 partitionValue: Partition,
@@ -1392,7 +1430,6 @@ private class ObservableAsyncOpenStorage: ObservableObject {
             let bsonValue = AnyBSON(partitionValue)
             if storage.partitionValue != bsonValue {
                 storage.partitionValue = bsonValue
-                storage.asyncOpen()
             }
         }
 
@@ -1403,7 +1440,6 @@ private class ObservableAsyncOpenStorage: ObservableObject {
                 storage.partitionValue = partitionValue
             }
             storage.configuration = configuration
-            storage.asyncOpen()
         }
     }
 }